* Extracted note into its own module
